pyproject.toml
[tool.poetry]
name = "django-auth-prod"
version = "0.1.0"
description = ""
authors = ["Lee-Roy King <leeroywking@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.8"
django = "^3.1.1"
gunicorn = "^20.0.4"
djangorestframework = "^3.11.1"
djangorestframework-simplejwt = "^4.4.0"
whitenoise = "^5.2.0"
django-cors-headers = "^3.5.0"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
